moved query to a query folder
